backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pdfplumber>=0.11.0
python-docx==1.1.0
google-genai>=1.0.0
pydantic==2.12.5
python-dotenv>=1.0.1
